+++ tests/Unit/Application/Contracts/Bitrix24Partners/Repository/InMemoryBitrix24PartnerRepositoryImplementation.php
@@ -134,23 +134,6 @@ class InMemoryBitrix24PartnerRepositoryImplementation implements Bitrix24Partner
     }
 
     #[\Override]
-    public function delete(Uuid $uuid): void
-    {
-        $this->logger->debug('b24PartnerRepository.delete', ['id' => $uuid->toRfc4122()]);
-
-        $bitrix24Partner = $this->getById($uuid);
-        if (Bitrix24PartnerStatus::deleted !== $bitrix24Partner->getStatus()) {
-            throw new InvalidArgumentException(sprintf(
-                'you cannot delete bitrix24 partner item «%s», they must be in status deleted, current status «%s»',
-                $bitrix24Partner->getId()->toRfc4122(),
-                $bitrix24Partner->getStatus()->name
-            ));
-        }
-
-        unset($this->items[$uuid->toRfc4122()]);
-    }
-
-    #[\Override]
     public function getById(Uuid $uuid): Bitrix24PartnerInterface
     {
         $this->logger->debug('b24PartnerRepository.getById', ['id' => $uuid->toRfc4122()]);